fix: Merge `Hide cast button patch`
