+++ src/main/resources/addresources/values/strings.xml
@@ -441,11 +441,6 @@ This is because Crowdin requires temporarily flattening this file and removing t
             <string name="revanced_hide_captions_button_summary_on">Captions button is hidden</string>
             <string name="revanced_hide_captions_button_summary_off">Captions button is shown</string>
         </patch>
-        <patch id="layout.buttons.cast.HideCastButtonPatch">
-            <string name="revanced_hide_cast_button_title">Hide cast button</string>
-            <string name="revanced_hide_cast_button_summary_on">Cast button is hidden</string>
-            <string name="revanced_hide_cast_button_summary_off">Cast button is shown</string>
-        </patch>
         <patch id="layout.buttons.navigation.NavigationButtonsPatch">
             <string name="revanced_navigation_buttons_screen_title">Navigation buttons</string>
             <string name="revanced_navigation_buttons_screen_summary">Hide or change buttons in the navigation bar</string>
@@ -522,6 +517,9 @@ This is because Crowdin requires temporarily flattening this file and removing t
             <string name="revanced_hide_player_buttons_title">Hide previous &amp; next video buttons</string>
             <string name="revanced_hide_player_buttons_summary_on">Buttons are hidden</string>
             <string name="revanced_hide_player_buttons_summary_off">Buttons are shown</string>
+            <string name="revanced_hide_cast_button_title">Hide cast button</string>
+            <string name="revanced_hide_cast_button_summary_on">Cast button is hidden</string>
+            <string name="revanced_hide_cast_button_summary_off">Cast button is shown</string>
         </patch>
         <patch id="layout.hide.albumcards.AlbumCardsResourcePatch">
             <string name="revanced_hide_album_cards_title">Hide album cards</string>
